unix/linux/perf: new package
This change adds a perf API client package, per golang/go#24918.

The package is created at import path golang.org/x/sys/unix/linux/perf
rather than golang.org/x/sys/linux/perf (as suggested in the issue),
because a unix/linux directory has been created in the sys repository
in the meantime, and it seems sensible to use it.

This CL implements the minimal functionality needed in order to start
using perf natively from Go with working code, and exploring what an
idiomatic Go API would look like.

A large portion of the perf API surface is covered, and seems to be
working. However, there is still much work to be done in terms of
figuring out API details, adding nice-to-have features, and testing
the code more in-depth.

README.md describes the current state of the package, as well as
the design of the tests. It also documents the kernel and system
configuration requirements for effective testing of package perf.

perf.go contains the central Event and Attr data structures,
definitions of event type constants, common operations on events,
and utility code used by both counting and sampling events.

count.go implements the counting side of the API.

record.go implements the sampling side of the API. It contains an
implementation of a context-aware ring buffer poller, as well as data
structures and decoders for all overflow record types.

group.go contains the Group API, offered because wiring events together
manually can be tiresome, and measuring multiple events together in
order to compare them is a fundamental M.O. of perf.

Finally, doc.go contains a few short code snippets, demonstrating
typical use of the package.

Package perf does not yet support reading perf.data files produced by
the Linux perf tool, in order to keep the scope of this change as small
as possible. Instead, this CL focuses only on the perf API itself.

Fixes golang/go#24918

Change-Id: I95c44479c2fa21ef5030c7d3b4c16a123b6871ae